requirements.txt
numpy
pillow
mss
keyboard
tensorflow
keras
scikit-learn
opencv-python
seaborn